fix-bugs-at-newsServices-delete-request

+++ client/src/services/newsServices.js
@@ -27,7 +27,7 @@ export const getAllNews = async () => {
 };
 
 //Método DELETE
-export const deletePost = async () => {
+export const deletePost = async (id) => {
     try{
         const token = localStorage.getItem('token');
         if (!token) {
@@ -37,7 +37,10 @@ export const deletePost = async () => {
             'Authorization': `Bearer ${token}`
         };  
 
+    if(confirm('¿Estás seguro de que quieres eliminar esta noticia?')=== true){
         const news =await axios.delete(`${url}/${id}`, {headers})
+        return news
+    }
     }
     catch(error){
         console.error('Error delete news:', error);
